+++ library/src/test/java/com/mux/player/CacheDatastoreTests.kt
@@ -2,9 +2,6 @@ package com.mux.player
 
 import android.content.Context
 import com.mux.player.internal.cache.CacheDatastore
-import com.mux.player.internal.cache.IndexSchema
-import com.mux.player.internal.cache.FileRecord
-import com.mux.player.internal.cache.toContentValues
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.Assert
@@ -12,7 +9,7 @@ import org.junit.Before
 import org.junit.Test
 import java.net.URL
 
-class CacheDatastoreTests: AbsRobolectricTest() {
+class CacheDatastoreTests : AbsRobolectricTest() {
 
   private lateinit var cacheDatastore: CacheDatastore
 
@@ -61,59 +58,6 @@ class CacheDatastoreTests: AbsRobolectricTest() {
     Assert.assertEquals(
       "cache key should be constructed properly",
       expectedKey, key
-    )
-  }
-
-  @Test
-  fun `FileRecord-toContentValues generates correct content values for its fields`() {
-    val record = FileRecord(
-      url = "url",
-      etag = "etag",
-      relativePath = "cacheFile",
-      lookupKey = "lookupKey",
-      downloadedAtUtcSecs = 1L,
-      cacheMaxAge = 2L,
-      resourceAge = 3L,
-      cacheControl = "cacheControl",
-      lastAccessUtcSecs = 4L
-    )
-
-    val contentValues = record.toContentValues()
-    Assert.assertEquals(
-      "last-access should be saved",
-      4L, contentValues.getAsLong(IndexSchema.FilesTable.Columns.lastAccessUnixTime)
-    )
-    Assert.assertEquals(
-      "url should be saved",
-      "url", contentValues.getAsString(IndexSchema.FilesTable.Columns.remoteUrl)
-    )
-    Assert.assertEquals(
-      "etag should be saved",
-      "etag", contentValues.getAsString(IndexSchema.FilesTable.Columns.etag)
-    )
-    Assert.assertEquals(
-      "file should be saved",
-      "cacheFile", contentValues.getAsString(IndexSchema.FilesTable.Columns.filePath)
-    )
-    Assert.assertEquals(
-      "lookup key should be saved",
-      "lookupKey", contentValues.getAsString(IndexSchema.FilesTable.Columns.lookupKey)
-    )
-    Assert.assertEquals(
-      "downloadedAt should be saved",
-      1L, contentValues.getAsLong(IndexSchema.FilesTable.Columns.downloadedAtUnixTime)
-    )
-    Assert.assertEquals(
-      "cacheMaxAge should be saved",
-      2L, contentValues.getAsLong(IndexSchema.FilesTable.Columns.maxAgeUnixTime)
-    )
-    Assert.assertEquals(
-      "age should be saved",
-      3L, contentValues.getAsLong(IndexSchema.FilesTable.Columns.resourceAgeUnixTime)
-    )
-    Assert.assertEquals(
-      "cache-control should be saved for future reference",
-      "cacheControl", contentValues.getAsString(IndexSchema.FilesTable.Columns.cacheControl)
     )
   }
 }
